Setup game repository

+++ composeApp/src/commonMain/kotlin/com/alextos/thousand/data/mappers/DiceRollMapper.kt
@@ -5,8 +5,8 @@ import com.alextos.thousand.data.models.combined.DiceRollWithDice
 import com.alextos.thousand.domain.models.DiceRoll
 
 fun DiceRoll.toEntity(
-    userId: Int,
-    turnId: Int,
+    userId: Long,
+    turnId: Long,
 ): DiceRollEntity = DiceRollEntity(
     id = id,
     userId = userId,
